+++ app/services/invitations/send_sms.rb
@@ -1,6 +1,7 @@
 module Invitations
   class SendSms < BaseService
     include Rails.application.routes.url_helpers
+    include Invitations::SmsContent
 
     def initialize(invitation:)
       @invitation = invitation
@@ -31,33 +32,11 @@ module Invitations
     end
 
     def content
-      send(:"content_for_#{@invitation.motif_category}")
-    end
-
-    def content_for_rsa_orientation
-      "#{applicant.full_name},\nVous êtes bénéficiaire du RSA et vous devez vous présenter à un rendez-vous " \
-        "d'orientation. Pour choisir la date et l'horaire de votre premier RDV, " \
-        "cliquez sur le lien suivant dans les #{number_of_days_to_accept_invitation} jours: " \
-        "#{redirect_invitations_url(params: { token: @invitation.token }, host: ENV['HOST'])}\n" \
-        "Ce rendez-vous est obligatoire. En cas de problème technique, contactez le "\
-        "#{@invitation.help_phone_number}."
-    end
-
-    def content_for_rsa_accompagnement
-      "#{applicant.full_name},\nVous êtes bénéficiaire du RSA et vous devez vous présenter à un rendez-vous " \
-        "d'accompagnement. Pour choisir la date et l'horaire de votre premier RDV, " \
-        "cliquez sur le lien suivant dans les #{number_of_days_to_accept_invitation} jours: " \
-        "#{redirect_invitations_url(params: { token: @invitation.token }, host: ENV['HOST'])}\n" \
-        "Ce rendez-vous est obligatoire. En l’absence d'action de votre part, " \
-        "le versement de votre RSA pourra être suspendu. En cas de problème technique, contactez le "\
-        "#{@invitation.help_phone_number}."
-    end
-
-    def content_for_rsa_orientation_on_phone_platform
-      "#{applicant.full_name},\nVous êtes bénéficiaire du RSA et vous devez contacter la plateforme départementale " \
-        "afin de démarrer votre parcours d’accompagnement. Pour cela, merci d’appeler le " \
-        "#{@invitation.help_phone_number} dans un délai de #{number_of_days_to_accept_invitation} jours. "\
-        "Cet appel est nécessaire pour le traitement de votre dossier."
+      if @invitation.reminder?
+        send(:"content_for_#{@invitation.motif_category}_reminder")
+      else
+        send(:"content_for_#{@invitation.motif_category}")
+      end
     end
 
     def phone_number_formatted
